Opprydning i skjema
